+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/internal/factory/NamingStrategyWrapperFactory.java
@@ -2,8 +2,8 @@ package org.hibernate.tool.orm.jbt.internal.factory;
 
 import org.hibernate.cfg.NamingStrategy;
 import org.hibernate.tool.orm.jbt.api.wrp.NamingStrategyWrapper;
+import org.hibernate.tool.orm.jbt.internal.util.ReflectUtil;
 import org.hibernate.tool.orm.jbt.internal.wrp.AbstractWrapper;
-import org.hibernate.tool.orm.jbt.util.ReflectUtil;
 
 public class NamingStrategyWrapperFactory {
