+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/internal/factory/RevengStrategyWrapperFactory.java
@@ -8,8 +8,8 @@ import org.hibernate.tool.internal.reveng.strategy.DefaultStrategy;
 import org.hibernate.tool.orm.jbt.api.wrp.RevengSettingsWrapper;
 import org.hibernate.tool.orm.jbt.api.wrp.RevengStrategyWrapper;
 import org.hibernate.tool.orm.jbt.api.wrp.Wrapper;
+import org.hibernate.tool.orm.jbt.internal.util.ReflectUtil;
 import org.hibernate.tool.orm.jbt.internal.wrp.AbstractWrapper;
-import org.hibernate.tool.orm.jbt.util.ReflectUtil;
 
 public class RevengStrategyWrapperFactory {
